+++ lib/keycloak-admin/representation/member_representation.rb
@@ -1,0 +1,11 @@
+module KeycloakAdmin
+  class MemberRepresentation < UserRepresentation
+    attr_accessor :membership_type
+
+    def self.from_hash(hash)
+      member                 = super(hash)
+      member.membership_type = hash["membershipType"]
+      member
+    end
+  end
+end